+++ bench/irmin-pack/bench_common.mli
@@ -1,6 +1,17 @@
+val default_results_dir : string
+val prepare_results_dir : string -> unit
 val reporter : ?prefix:string -> unit -> Logs.reporter
 val reset_stats : unit -> unit
 val with_timer : (unit -> 'a Lwt.t) -> (float * 'a) Lwt.t
+
+val with_progress_bar :
+  message:string ->
+  n:int ->
+  unit:string ->
+  sampling_interval:int ->
+  ((int64 -> unit) -> 'a) ->
+  'a
+
 val info : unit -> Irmin.Info.t
 val random_blob : unit -> string